fix(etl): cast numpy floats to native Python types for DB inserts
Aggregation and correlation batch inserts silently failed because
round() on pandas/numpy values returns np.float64, which psycopg2
serializes as literal "np.float64(...)" in SQL — causing PostgreSQL
to reject with "schema np does not exist". Wrap all numeric values
with float()/int() before insertion.

Fixes: sentiment_aggregations and sentiment_price_correlation tables
not being populated during backtest pipeline.

+++ apps/etl/aggregation_engine.py
@@ -98,9 +98,9 @@ class SentimentAggregator:
                         symbol,
                         bucket_time,
                         interval_label,
-                        round(avg_sentiment, 6),
-                        round(weighted_avg, 6),
-                        msg_count,
+                        float(round(avg_sentiment, 6)),
+                        float(round(weighted_avg, 6)),
+                        int(msg_count),
                         backtest_id,
                         None  # session_id (backtest mode)
                     ))
